<glide> Combine RequestOptions and BaseRequestOptions.
Instead of asking users to write their own BaseRequestOptions sublcass, we're now just generating a subclass and asking users to write GlideExtensions that are included in the generated subclass. It's confusing to have an otherwise unused base class and there's not too much overhead in generating a class that overrides a bunch of builder methods.